// src/app/model/empleados.ts
export interface Empleados {
    id?: number;
    fotografia?: string;
    nombre?: string;
    apellido?: string;
    puestoId?: number;
    fechaNacimiento?: string;
    fechaContratacion?: string;
    direccion?: string;
    telefono?: string;
    correoElectronico?: string;
    estadoId?: number;
    estado?: string;
    puesto?: string;
}